removing assemblies
